Fix creating a max-merge cache from a converted image.
When we have converted an image and then create a max-merge cache, we
end up with a girder item named "multi-source2.json".  When we ask to
create the max-merge cache, by the time the multi-source max-merge cache
gets processed, the name of the file ends up ending in json, and
large-image prefers the alphabetically first openslide source over the
tiff source, which fails to read all of the channels and other frames.
In this case, require that we use the tiff source for reading the
converted file.

Resolves #950