Fix the unroll checkbox disappearing
When the unroll checkbox was selected, it changes the range of the variable (XY, Z, T) to just have the one value, so that caused the slider and the checkbox to disappear. Fixed so that the checkbox is visible if there is an unroll, along with an "XY is unrolled" note.