Fix model inheritance pattern
Before, we had a collision, because ProxiedModel derived from CustomNimbusImageModel which derived from Model, and then we composed with AccessControlledModel. This led to potential issues before because the .load function would derive from Model and not AccessControlledModel, and so we had to switch the order of composition. Instead, now we just derive from AccessControlledModel as the base, thereby avoiding such conflicts.